+++ docs/CHANGELOG.md
@@ -1,5 +1,8 @@
 # Changelog
 
+### 2023-03-02
+* `depth` texture format
+
 ### 2023-03-01
 * Mesh rows alternate diagonal direction (see [MeshGrid](https://google.github.io/swissgl/#MeshGrid) example)
 * [wireframe()](https://github.com/google/swissgl/blob/8cf8cac20c4ec3352fec639c8d22dc5814d5e674/swissgl.js#L201) helper